remove prebuilt 2/N

+++ vulcast-rtc-sys/src/wrapper.hpp
@@ -10,7 +10,7 @@ class Producer;
 } // namespace mediasoupclient
 
 // foreign callback requesting frame in RGB (little-endian) format
-typedef void (*frame_callback_t)(const void *ctx, size_t width, size_t height,
+typedef void (*frame_callback_t)(const void *ctx, uint32_t width, uint32_t height,
                                  int64_t timestamp, uint8_t *);
 
 struct SignalHandler {
@@ -57,8 +57,8 @@ mediasoupclient::Producer *producer_new_from_fake_audio(Broadcaster *b);
 mediasoupclient::Producer *producer_new_from_fake_video(Broadcaster *b);
 mediasoupclient::Producer *producer_new_from_vcm_capturer(Broadcaster *b);
 mediasoupclient::Producer *
-producer_new_from_foreign(Broadcaster *b, size_t width, size_t height,
-                          size_t fps, void *ctx, frame_callback_t callback);
+producer_new_from_foreign(Broadcaster *b, uint32_t width, uint32_t height,
+                          uint32_t fps, void *ctx, frame_callback_t callback);
 void producer_delete(mediasoupclient::Producer *producer);
 
 void debug_enumerate_capture_devices();
